test DrawTool
