+++ src/map/tool/DrawTool.js
@@ -3,18 +3,9 @@ import { isNil } from '../../core/util';
 import { extendSymbol } from '../../core/util/style';
 import { getExternalResources } from '../../core/util/resource';
 import { stopPropagation } from '../../core/util/dom';
-import Coordinate from '../../geo/Coordinate';
-import Point from '../../geo/Point';
-import Marker from '../../geometry/Marker';
-import Polygon from '../../geometry/Polygon';
-import LineString from '../../geometry/LineString';
-import Circle from '../../geometry/Circle';
-import Ellipse from '../../geometry/Ellipse';
-import ArcCurve from '../../geometry/ArcCurve';
-import CubicBezierCurve from '../../geometry/CubicBezierCurve';
-import QuadBezierCurve from '../../geometry/QuadBezierCurve';
 import VectorLayer from '../../layer/VectorLayer';
 import MapTool from './MapTool';
+import RegisterModes from './DrawToolGeometry';
 
 /**
  * @property {Object} [options=null] - construct options
@@ -81,6 +72,18 @@ class DrawTool extends MapTool {
     }
 
     /**
+     * Register modes for DrawTool
+     * @param modes
+     */
+    static registeredModes(modes) {
+        if (modes) {
+            for (const key of Reflect.ownKeys(modes)) {
+                DrawTool.registerMode(key, modes[key]);
+            }
+        }
+    }
+
+    /**
      * Get mode actions by mode name
      * @param  {String} name DrawTool mode name
      * @return {Object}      mode actions
@@ -101,6 +104,19 @@ class DrawTool extends MapTool {
     constructor(options) {
         super(options);
         this._checkMode();
+
+        /**
+         * events
+         * @type {{click: *, mousemove: *, dblclick: *, mousedown: *, mouseup: *, drag: *}}
+         */
+        this.events = {
+            'click': this._firstClickHandler,
+            'mousemove': this._mouseMoveHandler,
+            'dblclick': this._doubleClickHandler,
+            'mousedown': this._mouseDownHandler,
+            'mouseup': this._mouseUpHandler,
+            'drag': this._mouseMoveHandler
+        };
     }
 
     /**
@@ -281,20 +297,16 @@ class DrawTool extends MapTool {
 
     getEvents() {
         const action = this._getRegisterMode()['action'];
-        if (action === 'clickDblclick') {
-            return {
-                'click': this._clickForPath,
-                'mousemove': this._mousemoveForPath,
-                'dblclick': this._dblclickForPath
-            };
-        } else if (action === 'click') {
-            return {
-                'click': this._clickForPoint
-            };
-        } else if (action === 'drag') {
-            return {
-                'mousedown': this._mousedownToDraw
-            };
+        const _events = {};
+        if (Array.isArray(action)) {
+            for (let i = 0; i < action.length; i++) {
+                if (action[i] === 'drag') {
+                    _events['mousemove'] = this.events[action[i]];
+                } else {
+                    _events[action[i]] = this.events[action[i]];
+                }
+            }
+            return _events;
         }
         return null;
     }
@@ -304,186 +316,121 @@ class DrawTool extends MapTool {
         drawLayer.addGeometry(geometry);
     }
 
-    _clickForPoint(param) {
-        const registerMode = this._getRegisterMode();
-        this._geometry = registerMode['create'](param['coordinate'], param);
-        if (this.options['symbol'] && this.options.hasOwnProperty('symbol')) {
-            this._geometry.setSymbol(this.options['symbol']);
-        }
-        this.endDraw();
+    /**
+     * mouse down start draw
+     * @param event
+     * @private
+     */
+    _mouseDownHandler(event) {
+        this._createGeometry(event);
     }
 
-    _clickForPath(param) {
+    /**
+     * handle mouse up event
+     * @param event
+     * @private
+     */
+    _mouseUpHandler(event) {
+        this.endDraw(event);
+    }
+
+    /**
+     * handle mouse first click handle
+     * @param event
+     * @private
+     */
+    _firstClickHandler(event) {
         const registerMode = this._getRegisterMode();
-        const coordinate = param['coordinate'];
-        const symbol = this.getSymbol();
+        const coordinate = event['coordinate'];
         if (!this._geometry) {
-            this._clickCoords = [coordinate];
-            this._geometry = registerMode['create'](this._clickCoords, param);
-            if (symbol) {
-                this._geometry.setSymbol(symbol);
-            }
-            this._addGeometryToStage(this._geometry);
-            /**
-             * drawstart event.
-             *
-             * @event DrawTool#drawstart
-             * @type {Object}
-             * @property {String} type - drawstart
-             * @property {DrawTool} target - draw tool
-             * @property {Coordinate} coordinate - coordinate of the event
-             * @property {Point} containerPoint  - container point of the event
-             * @property {Point} viewPoint       - view point of the event
-             * @property {Event} domEvent                 - dom event
-             */
-            this._fireEvent('drawstart', param);
+            this._createGeometry(event);
         } else {
-            if (!isNil(this._historyPointer)) {
+            if (this._clickCoords.length > 0 &&
+                this.getMap().computeLength(coordinate, this._clickCoords[this._clickCoords.length - 1]) < 0.01) {
+                return;
+            }
+            if (!(this._historyPointer === null)) {
                 this._clickCoords = this._clickCoords.slice(0, this._historyPointer);
             }
             this._clickCoords.push(coordinate);
             this._historyPointer = this._clickCoords.length;
-            registerMode['update'](this._clickCoords, this._geometry, param);
-            /**
-             * drawvertex event.
-             *
-             * @event DrawTool#drawvertex
-             * @type {Object}
-             * @property {String} type - drawvertex
-             * @property {DrawTool} target - draw tool
-             * @property {Geometry} geometry - geometry drawn
-             * @property {Coordinate} coordinate - coordinate of the event
-             * @property {Point} containerPoint  - container point of the event
-             * @property {Point} viewPoint       - view point of the event
-             * @property {Event} domEvent                 - dom event
-             */
-            this._fireEvent('drawvertex', param);
-
+            if (registerMode['limitClickCount'] && registerMode['limitClickCount'] === this._historyPointer) {
+                registerMode['update'](this._clickCoords, this._geometry, event);
+                this.endDraw(event);
+            } else {
+                registerMode['update'](this._clickCoords, this._geometry, event);
+            }
+            this._fireEvent('drawvertex', event);
+        }
+        if (this.getMode() === 'point') {
+            this.endDraw(event);
         }
     }
 
-    _mousemoveForPath(param) {
+    /**
+     * 第一次事件创建相关geometry
+     * @param event
+     * @private
+     */
+    _createGeometry(event) {
+        const registerMode = this._getRegisterMode();
+        const coordinate = event['coordinate'];
+        const symbol = this.getSymbol();
+        if (!this._geometry) {
+            this._clickCoords = [coordinate];
+            this._geometry = registerMode['create'](this._clickCoords, event);
+            if (symbol) {
+                this._geometry.setSymbol(symbol);
+            }
+            this._addGeometryToStage(this._geometry);
+            this._fireEvent('drawstart', event);
+        }
+    }
+
+    /**
+     * handle mouse move event
+     * FIXME 当为freehand模式时，鼠标按下松开而不移动会造成构造的geometry不完整
+     * @param event
+     * @private
+     */
+    _mouseMoveHandler(event) {
         const map = this.getMap();
+        const coordinate = event['coordinate'];
         if (!this._geometry || !map || map.isInteracting()) {
             return;
         }
-        const containerPoint = this._getMouseContainerPoint(param);
+        if (map.computeLength(coordinate, this._clickCoords[this._clickCoords.length - 1]) < 0.01) {
+            return;
+        }
+        const containerPoint = this._getMouseContainerPoint(event);
         if (!this._isValidContainerPoint(containerPoint)) {
             return;
         }
-        const coordinate = param['coordinate'];
         const registerMode = this._getRegisterMode();
         const path = this._clickCoords.slice(0, this._historyPointer);
         if (path && path.length > 0 && coordinate.equals(path[path.length - 1])) {
             return;
         }
-        registerMode['update'](path.concat([coordinate]), this._geometry, param);
-        /**
-         * mousemove event.
-         *
-         * @event DrawTool#mousemove
-         * @type {Object}
-         * @property {String} type - mousemove
-         * @property {DrawTool} target - draw tool
-         * @property {Geometry} geometry - geometry drawn
-         * @property {Coordinate} coordinate - coordinate of the event
-         * @property {Point} containerPoint  - container point of the event
-         * @property {Point} viewPoint       - view point of the event
-         * @property {Event} domEvent                 - dom event
-         */
-        this._fireEvent('mousemove', param);
+        if (!registerMode.freehand) {
+            registerMode['update'](path.concat([coordinate]), this._geometry, event);
+        } else {
+            if (!(this._historyPointer === null)) {
+                this._clickCoords = this._clickCoords.slice(0, this._historyPointer);
+            }
+            this._clickCoords.push(coordinate);
+            this._historyPointer = this._clickCoords.length;
+            registerMode['update'](this._clickCoords, this._geometry, event);
+        }
+        this._fireEvent('mousemove', event);
     }
 
-    _dblclickForPath(param) {
-        if (!this._geometry) {
-            return;
-        }
-        const containerPoint = this._getMouseContainerPoint(param);
-        if (!this._isValidContainerPoint(containerPoint)) {
-            return;
-        }
-        const registerMode = this._getRegisterMode();
-        const path = this._clickCoords;
-        if (path.length < 2) {
-            return;
-        }
-        //remove duplicate vertexes
-        const nIndexes = [];
-        for (let i = 1, len = path.length; i < len; i++) {
-            if (path[i].x === path[i - 1].x && path[i].y === path[i - 1].y) {
-                nIndexes.push(i);
-            }
-        }
-        for (let i = nIndexes.length - 1; i >= 0; i--) {
-            path.splice(nIndexes[i], 1);
-        }
-
-        if (path.length < 2 || (this._geometry && (this._geometry instanceof Polygon) && path.length < 3)) {
-            return;
-        }
-        registerMode['update'](path, this._geometry, param);
-        this.endDraw(param);
-    }
-
-    _mousedownToDraw(param) {
-        const map = this._map;
-        const registerMode = this._getRegisterMode();
-        const me = this,
-            firstPoint = this._getMouseContainerPoint(param);
-        if (!this._isValidContainerPoint(firstPoint)) {
-            return false;
-        }
-
-        function genGeometry(evt) {
-            const symbol = me.getSymbol();
-            let geometry = me._geometry;
-            if (!geometry) {
-                geometry = registerMode['create'](evt.coordinate, evt);
-                geometry.setSymbol(symbol);
-                me._addGeometryToStage(geometry);
-                me._geometry = geometry;
-            } else {
-                registerMode['update'](evt.coordinate, geometry, evt);
-            }
-        }
-
-        function onMouseMove(evt) {
-            if (!this._geometry) {
-                return false;
-            }
-            const current = this._getMouseContainerPoint(evt);
-            if (!this._isValidContainerPoint(current)) {
-                return false;
-            }
-            genGeometry(evt);
-            this._fireEvent('mousemove', param);
-            return false;
-        }
-        const onMouseUp = function (evt) {
-            map.off('mousemove', onMouseMove, this);
-            map.off('mouseup', onMouseUp, this);
-            if (!this.options['ignoreMouseleave']) {
-                map.off('mouseleave', onMouseUp, this);
-            }
-            if (!this._geometry) {
-                return false;
-            }
-            const current = this._getMouseContainerPoint(evt);
-            if (this._isValidContainerPoint(current)) {
-                genGeometry(evt);
-            }
-            this.endDraw(param);
-            return false;
-        };
-
-        this._fireEvent('drawstart', param);
-        genGeometry(param);
-        map.on('mousemove', onMouseMove, this);
-        map.on('mouseup', onMouseUp, this);
-        if (!this.options['ignoreMouseleave']) {
-            map.on('mouseleave', onMouseUp, this);
-        }
-        return false;
+    /**
+     * handle mouse double click event
+     * @param event
+     * @private
+     */
+    _doubleClickHandler(event) {
+        this.endDraw(event);
     }
 
     /**
@@ -580,188 +527,6 @@ class DrawTool extends MapTool {
 
 DrawTool.mergeOptions(options);
 
-DrawTool.registerMode('circle', {
-    'action': 'drag',
-    'create': function (coordinate) {
-        return new Circle(coordinate, 0);
-    },
-    'update': function (coordinate, geometry) {
-        const map = geometry.getMap();
-        const radius = map.computeLength(geometry.getCenter(), coordinate);
-        geometry.setRadius(radius);
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
-
-DrawTool.registerMode('ellipse', {
-    'action': 'drag',
-    'create': function (coordinate) {
-        return new Ellipse(coordinate, 0, 0);
-    },
-    'update': function (coordinate, geometry) {
-        const map = geometry.getMap();
-        const center = geometry.getCenter();
-        const rx = map.computeLength(center, new Coordinate({
-            x: coordinate.x,
-            y: center.y
-        }));
-        const ry = map.computeLength(center, new Coordinate({
-            x: center.x,
-            y: coordinate.y
-        }));
-        geometry.setWidth(rx * 2);
-        geometry.setHeight(ry * 2);
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
-
-DrawTool.registerMode('rectangle', {
-    'action': 'drag',
-    'create': function (coordinate, param) {
-        const rect = new Polygon([]);
-        rect._firstClick = param['containerPoint'];
-        return rect;
-    },
-    'update': function (coordinate, geometry, param) {
-        const map = geometry.getMap();
-        const containerPoint = param['containerPoint'];
-        const firstClick = geometry._firstClick;
-        const ring = [
-            [firstClick.x, firstClick.y],
-            [containerPoint.x, firstClick.y],
-            [containerPoint.x, containerPoint.y],
-            [firstClick.x, containerPoint.y],
-        ];
-        geometry.setCoordinates(ring.map(c => map.containerPointToCoord(new Point(c))));
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
-
-DrawTool.registerMode('point', {
-    'action': 'click',
-    'create': function (coordinate) {
-        return new Marker(coordinate);
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
-
-DrawTool.registerMode('polygon', {
-    'action': 'clickDblclick',
-    'create': function (path) {
-        return new LineString(path);
-    },
-    'update': function (path, geometry) {
-        const symbol = geometry.getSymbol();
-        geometry.setCoordinates(path);
-
-        const layer = geometry.getLayer();
-        if (layer) {
-            let polygon = layer.getGeometryById('polygon');
-            if (!polygon && path.length >= 3) {
-                polygon = new Polygon([path], {
-                    'id': 'polygon'
-                });
-                if (symbol) {
-                    const pSymbol = extendSymbol(symbol, {
-                        'lineOpacity': 0
-                    });
-                    polygon.setSymbol(pSymbol);
-                }
-                polygon.addTo(layer);
-            }
-            if (polygon) {
-                polygon.setCoordinates(path);
-            }
-        }
-    },
-    'generate': function (geometry) {
-        return new Polygon(geometry.getCoordinates(), {
-            'symbol': geometry.getSymbol()
-        });
-    }
-});
-
-DrawTool.registerMode('linestring', {
-    'action': 'clickDblclick',
-    'create': function (path) {
-        return new LineString(path);
-    },
-    'update': function (path, geometry) {
-        geometry.setCoordinates(path);
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
-
-DrawTool.registerMode('arccurve', {
-    'action': 'clickDblclick',
-    'create': function (path) {
-        return new ArcCurve(path);
-    },
-    'update': function (path, geometry) {
-        geometry.setCoordinates(path);
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
-
-DrawTool.registerMode('quadbeziercurve', {
-    'action': 'clickDblclick',
-    'create': function (path) {
-        return new QuadBezierCurve(path);
-    },
-    'update': function (path, geometry) {
-        geometry.setCoordinates(path);
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
-
-DrawTool.registerMode('cubicbeziercurve', {
-    'action': 'clickDblclick',
-    'create': function (path) {
-        return new CubicBezierCurve(path);
-    },
-    'update': function (path, geometry) {
-        geometry.setCoordinates(path);
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
-
-DrawTool.registerMode('boxZoom', {
-    'action': 'drag',
-    'create': function (coordinate) {
-        const marker = new Marker(coordinate);
-        marker._firstClick = coordinate;
-        return marker;
-    },
-    'update': function (coordinate, geometry, param) {
-        const map = geometry.getMap();
-        const p1 = map.coordToContainerPoint(geometry._firstClick),
-            p2 = param['containerPoint'];
-        const coord = map.containerPointToCoordinate(new Coordinate(Math.min(p1.x, p2.x), Math.min(p1.y, p2.y)));
-        geometry.setCoordinates(coord)
-            .updateSymbol({
-                markerWidth  : Math.abs(p1.x - p2.x),
-                markerHeight : Math.abs(p1.y - p2.y)
-            });
-    },
-    'generate': function (geometry) {
-        return geometry;
-    }
-});
+DrawTool.registeredModes(RegisterModes);
 
 export default DrawTool;
